+++ src/Webhook/WebhookController.php
@@ -16,7 +16,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Coincharge\Shopware\Webhook\WebhookServiceInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Shopware\Core\Framework\Context;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\Response;
 
 #[Route(defaults: ['_routeScope' => ['api']])]
@@ -28,7 +28,7 @@ class WebhookController extends AbstractController
     {
         $this->webhookRouter = $webhookRouter;
     }
-    #[Route(path: '/api/_action/coincharge/webhook-endpoint', name: 'api.action.coincharge.webhook.endpoint', methods: ['POST'], defaults: ['XmlHttpRequest' => true, 'auth_required' => false, 'csrf_protected' => false])]
+    #[Route(path: '/api/_action/coincharge/webhook-endpoint', name: 'api.action.coincharge.webhook.endpoint', methods: ['POST'], defaults: ['XmlHttpRequest' => true, 'auth_required' => false])]
     public function endpoint(Request $request, Context $context): Response
     {
         return $this->webhookRouter->route($request, $context);